# [V1][Perf] Reduce scheduling overhead in model runner after cuda sync