We do some runner bookkeeping CPU operations after decoding iteration. We could parallelize some bookkeeping work while waiting on cuda sync. After the cuda sync, we only need to do simple and fast updates.

The change should reduce scheduling overhead between decode iterations by ~20%. (See attached gpu trace)

Before the optimization,
<img width="1512" alt="Screenshot 2025-01-15 at 10 39 51 AM" src="https://github.com/user-attachments/assets/5aac1d4a-6970-45e2-8889-1bbefa2ed57c" />

After the optimization,
<img width="1501" alt="Screenshot 2025-01-15 at 10 40 42 AM" src="https://github.com/user-attachments/assets/b41829ce-f81b-4d22-858a-74cb8ee66290" />


E2E latency benchmark, ran

VLLM_USE_V1=1 python3 benchmarks/benchmark_latency.py --model "/data/users/ktong/llama/llm_8b_oss" --tensor-parallel-size 1 --input_len 1000 --batch_size 32  

Output (1-2% e2e latency reduction):
Avg latency: 2.338167402730323 seconds
10% percentile latency: 2.3207896508742123 seconds
25% percentile latency: 2.3264574960339814 seconds
50% percentile latency: 2.3333765944698825 seconds
75% percentile latency: 2.343035737867467 seconds
90% percentile latency: 2.3567665563430635 seconds
99% percentile latency: 2.3934816433605737 seconds