@robertgshaw2-redhat 2025-01-16
Wow, great idea. Im going to run some perfomance analysis on this tomorrow.