@WoosukKwon 2025-01-26
@youngkent Thanks for the PR! This change helps vLLM's performance noticeably.